codigos/requirements.txt
absl-py==2.3.0
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.4
jax==0.6.1
jaxlib==0.6.1
kiwisolver==1.4.8
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
